Enable ignorewarningsforurls for AnchorCheck
Some websites, such as GitHub, generate some of the anchors in
JavaScript. As a result, AnchorCheck tends to warn of missing anchors.

This PR introduces the `url-anchorcheck-anchor-not-found` warning and
changes AnchorCheck so it tags all its warnings. This enables the user
to use the recently introduced `ignorewarningsforurls` feature to
silence AnchorCheck for selected URLs.